Move PrivacyManagerTest to PrivacyManager plugin as DataPurgingTest.php.
